+++ tests/testthat/test-BFBayesFactor.R
@@ -75,4 +75,21 @@ if (requireNamespace("BayesFactor", quietly = TRUE)) {
   # test_that("p_direction", {
   #   expect_equal(as.numeric(p_direction(x)), 91.9, tol=0.1)
   # })
+
+
+  test_that("rope_range", {
+    x <- BayesFactor::lmBF(len ~ supp + dose, data = ToothGrowth)
+    testthat::expect_equal(rope_range(x)[2], sd(ToothGrowth$len)/10)
+
+    x <- BayesFactor::ttestBF(ToothGrowth$len[ToothGrowth$supp=="OJ"],
+                              ToothGrowth$len[ToothGrowth$supp=="VC"])
+    testthat::expect_equal(rope_range(x)[2], sd(ToothGrowth$len)/10)
+
+    # else
+    x <- BayesFactor::correlationBF(ToothGrowth$len, ToothGrowth$dose)
+    testthat::expect_equal(rope_range(x), c(-0.1, 0.1))
+  })
+
+
+
 }