+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Bug fixes
 
+- `rope_range()` for `BayesFactor` models returns non-`NA` values ( #343 )
 - `ci()` supports `emmeans` - both Bayesian and frequentist ( #312 - cross fix with `parameters`)
 
 # bayestestR 0.7.3